+++ SEProjectV2.2/src/database/allProducts/AllProductsInList.java
@@ -1,0 +1,12 @@
+package database.allProducts;
+
+public class AllProductsInList extends AllProducts{ // Class for List Output
+	
+	public AllProductsInList(){
+		super();
+	}
+
+	public void search(String toFind) {
+		//TODO
+	}
+}
