--- SEProjectV2.2/src/start/userInterface/changeCategory/ChangeCategory.java
@@ -1,12 +1,0 @@
-package start.userInterface.changeCategory;
-
-
-@SuppressWarnings("serial")
-public class ChangeCategory extends Interface{
-	
-	// view of the window
-	public ChangeCategory (){
-		super();
-	} 
-
-}
