+++ SEProjectV2.2/src/userInterface/changeCategory/ChangeCategory.java
@@ -1,0 +1,12 @@
+package userInterface.changeCategory;
+
+
+@SuppressWarnings("serial")
+public class ChangeCategory extends Interface{
+	
+	// view of the window
+	public ChangeCategory (){
+		super();
+	} 
+
+}
