+++ SEProjectV2.2/src/userInterface/mainWindow/MainWindow.java
@@ -1,0 +1,15 @@
+package userInterface.mainWindow;
+
+
+@SuppressWarnings("serial")
+public class MainWindow extends Interface {
+	
+	
+	
+	public MainWindow() {
+		super();
+		
+	}
+}
+
+
