--- SEProjectV2/src/start/Main.java
@@ -1,14 +1,0 @@
-package start;
-
-public class Main {
-
-	public static void main(String[] args) {
-		start.userInterface.mainWindow.MainWindow myMainWindow = new start.userInterface.mainWindow.MainWindow (); 
-		start.userInterface.productView.ProductView myProductView = new start.userInterface.productView.ProductView (); 
-		start.userInterface.addProductInterface.AddProduct myProduct1 = new start.userInterface.addProductInterface.AddProduct ();
-		start.userInterface.changeCategory.ChangeCategory myCategory1 = new start.userInterface.changeCategory.ChangeCategory ();
-		start.exeptions.Exception myException1 = new start.exeptions.Exception ();
-		System.out.println("Test bin ich da");
-	}
-
-}
