--- SEProjectV2/src/start/userInterface/addProductInterface/AddProduct.java
@@ -1,12 +1,0 @@
-package start.userInterface.addProductInterface;
-
-
-@SuppressWarnings("serial")
-public class AddProduct extends Interface {
-	
-	// view of the window
-	public AddProduct (){
-		super();	
-	} // addProduct	
-	
-} // public class addProduct
