--- SEProjectV2/src/start/userInterface/mainWindow/MainWindow.java
@@ -1,12 +1,0 @@
-package start.userInterface.mainWindow;
-
-
-@SuppressWarnings("serial")
-public class MainWindow extends Interface {
-	
-	public MainWindow() {
-		super();
-	}
-}
-
-
